add create designers and portfilios and merge with frontend

+++ server/src/routers/tags.js
@@ -4,7 +4,8 @@ const { Router } = require('express');
 const {
     getAll: getAll_,
     create: create_,
-    deleteTag: deleteTag_
+    deleteTag: deleteTag_,
+    updateTag: updateTag_
 } = require('../database/tags');
 
 async function getAll(req, res) {
@@ -18,26 +19,52 @@ async function getAll(req, res) {
     res.sendStatus(204);
 }
 
-async function create({ body: { name } }, res) {
-    let result = await create_(name);
+async function create({ body: { name }, session }, res) {
+    if (session.role !== undefined && session.role.indexOf('admin') !== -1) {
+        let result = await create_(name);
 
-    if(result.isSuccess){
-        res.json(result);
+        if (result.isSuccess) {
+            res.json(result);
+            return;
+        }
+
+        res.sendStatus(520);
         return;
     }
 
-    res.sendStatus(205)
+    res.sendStatus(401);
 }
 
-async function deleteTag({params: {id}}, res){
-    let result = await deleteTag_(id);
+async function deleteTag({ params: { id }, session }, res) {
+    if (session.role !== undefined && session.role.indexOf('admin') !== -1) {
+        let result = await deleteTag_(id);
 
-    if(result.isSuccess){
-        res.json(result);
+        if (result.isSuccess) {
+            res.json(result);
+            return;
+        }
+
+        res.sendStatus(520);
         return;
     }
 
-    res.sendStatus(204);
+    res.sendStatus(401);
+}
+
+async function updateTag({ params: { id }, body: { name }, session }, res) {
+    if (session.role !== undefined && session.role.indexOf('admin') !== -1) {
+        let result = await updateTag_(id, name);
+
+        if(result.isSuccess){
+            res.sendStatus(204);
+            return;
+        }
+
+        res.sendStatus(520);
+        return;
+    }
+
+    res.sendStatus(401);
 }
 
 function index() {
@@ -48,6 +75,8 @@ function index() {
     router.post('/', create);
 
     router.delete('/:id', deleteTag);
+
+    router.put('/:id', updateTag);
 
     return router;
 }
